Cart fixed

+++ MyAspNetMvcApp/Areas/OrderFramework/Models/OrderDetail.cs
@@ -15,7 +15,7 @@ namespace MyAspNetMvcApp.Areas.OrderFramework.Models
         public int OrderId { get; set; }
         public int ItemId { get; set; }
         public int Quantity { get; set; }
-        public double UnitPrice { get; set; }
+        public decimal UnitPrice { get; set; }
         public virtual Product Item { get; set; }
         public virtual Order Order { get; set; }
     }
